server/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
asyncpg==0.30.0
geoalchemy2==0.18.0
alembic==1.16.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic[email]==2.11.9
websockets==15.0.1
python-dotenv==1.1.1
requests==2.32.5
aiosqlite==0.20.0